requirements.txt
django
pillow
image
django-cleanup
django-crispy-forms
django-resized
django-filter
